seperate connection and logger
